+++ src/extensions/markdown/Table/index.ts
@@ -1,9 +1,10 @@
-import type {Action, ExtensionAuto} from '../../../core';
-import {goToNextCell} from '../../../table-utils';
+import type {Action, ExtensionAuto} from '#core';
+import {goToNextCell} from 'src/table-utils';
 
 import {TableSpecs} from './TableSpecs';
 import {createTableAction, deleteTableAction} from './actions/tableActions';
 import * as TableActions from './actions/tableActions';
+import {moveToNextRowCommand} from './commands';
 import * as TableHelpers from './helpers';
 import {tableCellContextPlugin} from './plugins/TableCellContextPlugin';
 
@@ -16,6 +17,8 @@ export const Table: ExtensionAuto = (builder) => {
     builder.addKeymap(() => ({
         Tab: goToNextCell('next'),
         'Shift-Tab': goToNextCell('prev'),
+        Enter: moveToNextRowCommand,
+        'Shift-Enter': moveToNextRowCommand,
     }));
 
     builder.addAction('createTable', createTableAction);