+++ src/markup/codemirror/search-plugin/plugin.ts
@@ -19,18 +19,17 @@ import {
     keymap,
 } from '@codemirror/view';
 
-import type {MarkdownEditorMode} from '../../../bundle';
-import type {EventMap} from '../../../bundle/Editor';
-import type {RendererItem} from '../../../extensions';
-import {debounce} from '../../../lodash';
-import type {Receiver} from '../../../utils';
+import type {MarkdownEditorMode} from 'src/bundle';
+import type {EventMap} from 'src/bundle/Editor';
+import type {RendererItem} from 'src/extensions';
+import {renderSearchPopup} from 'src/modules/search';
+import type {Receiver} from 'src/utils';
+
 import {ReactRendererFacet} from '../react-facet';
 
-import {renderSearchPopup} from './view/SearchPopup';
+import {searchTheme} from './theme';
 
 type SearchQueryConfig = ConstructorParameters<typeof SearchQuery>[0];
-
-const INPUT_DELAY = 200;
 
 export interface SearchPanelPluginParams {
     anchorSelector: string;
@@ -44,24 +43,20 @@ export const SearchPanelPlugin = (params: SearchPanelPluginParams) =>
             readonly view: EditorView;
             readonly params: SearchPanelPluginParams;
 
-            anchor: HTMLElement | null;
-            renderer: RendererItem | null;
-            searchConfig: SearchQueryConfig = {
-                search: '',
-                caseSensitive: false,
-                wholeWord: false,
-                replace: '',
-            };
+            renderer: RendererItem;
             receiver: Receiver<EventMap> | undefined;
 
-            setViewSearchWithDelay: (config: Partial<SearchQueryConfig>) => void;
+            panelOpened: boolean;
+            searchState: SearchQuery | null;
 
             constructor(view: EditorView) {
                 this.view = view;
-                this.anchor = null;
-                this.renderer = null;
                 this.params = params;
                 this.receiver = params.receiver;
+
+                this.panelOpened = searchPanelOpen(view.state);
+                this.searchState = getSearchQuery(view.state);
+                this.renderer = this.createRenderer();
 
                 this.handleClose = this.handleClose.bind(this);
                 this.handleChange = this.handleChange.bind(this);
@@ -69,60 +64,49 @@ export const SearchPanelPlugin = (params: SearchPanelPluginParams) =>
                 this.handleSearchPrev = this.handleSearchPrev.bind(this);
                 this.handleReplaceNext = this.handleReplaceNext.bind(this);
                 this.handleReplaceAll = this.handleReplaceAll.bind(this);
-                this.handleSearchConfigChange = this.handleSearchConfigChange.bind(this);
                 this.handleEditorModeChange = this.handleEditorModeChange.bind(this);
 
-                this.setViewSearchWithDelay = debounce(
-                    this.setViewSearch,
-                    this.params.inputDelay ?? INPUT_DELAY,
-                );
                 this.receiver?.on('change-editor-mode', this.handleEditorModeChange);
             }
 
             update(update: ViewUpdate): void {
                 const isPanelOpen = searchPanelOpen(update.state);
+                const searchQuery = getSearchQuery(update.state);
 
-                if (isPanelOpen && !this.renderer) {
-                    const initial = getSearchQuery(update.state);
-                    this.anchor = document.querySelector(this.params.anchorSelector);
-                    this.renderer = this.view.state
-                        .facet(ReactRendererFacet)
-                        .createItem('cm-search', () =>
-                            renderSearchPopup({
-                                initial,
-                                open: true,
-                                anchor: this.anchor,
-                                onChange: this.handleChange,
-                                onClose: this.handleClose,
-                                onSearchNext: this.handleSearchNext,
-                                onSearchPrev: this.handleSearchPrev,
-                                onReplaceNext: this.handleReplaceNext,
-                                onReplaceAll: this.handleReplaceAll,
-                                onConfigChange: this.handleSearchConfigChange,
-                            }),
-                        );
-                } else if (!isPanelOpen && this.renderer) {
-                    this.renderer?.remove();
-                    this.renderer = null;
+                if (isPanelOpen !== this.panelOpened || searchQuery !== this.searchState) {
+                    this.panelOpened = isPanelOpen;
+                    this.searchState = searchQuery;
+                    this.renderer.rerender();
                 }
             }
 
             destroy() {
-                this.renderer?.remove();
-                this.renderer = null;
+                this.renderer.remove();
                 this.receiver?.off('change-editor-mode', this.handleEditorModeChange);
             }
 
-            setViewSearch(config: Partial<SearchQueryConfig>) {
-                this.searchConfig = {
-                    ...this.searchConfig,
-                    ...config,
-                };
-                const searchQuery = new SearchQuery({
-                    ...this.searchConfig,
-                });
+            createRenderer() {
+                return this.view.state.facet(ReactRendererFacet).createItem('cm-search', () => {
+                    if (!this.panelOpened || !this.searchState) return null;
 
-                this.view.dispatch({effects: setSearchQuery.of(searchQuery)});
+                    const anchor = this.view.dom.ownerDocument.querySelector(
+                        this.params.anchorSelector,
+                    );
+
+                    if (!anchor) return null;
+
+                    return renderSearchPopup({
+                        open: true,
+                        anchor: anchor,
+                        state: this.searchState,
+                        onClose: this.handleClose,
+                        onChange: this.handleChange,
+                        onSearchNext: this.handleSearchNext,
+                        onSearchPrev: this.handleSearchPrev,
+                        onReplaceNext: this.handleReplaceNext,
+                        onReplaceAll: this.handleReplaceAll,
+                    });
+                });
             }
 
             handleEditorModeChange({mode}: {mode: MarkdownEditorMode}) {
@@ -131,13 +115,23 @@ export const SearchPanelPlugin = (params: SearchPanelPluginParams) =>
                 }
             }
 
-            handleChange(search: string) {
-                this.setViewSearchWithDelay({search});
+            handleChange(config: SearchQueryConfig) {
+                this.view.dispatch({
+                    effects: setSearchQuery.of(
+                        new SearchQuery({
+                            search: config.search,
+                            replace: config.replace,
+                            caseSensitive: config.caseSensitive,
+                            wholeWord: config.wholeWord,
+                        }),
+                    ),
+                });
             }
 
             handleClose() {
-                this.setViewSearch({search: ''});
+                this.handleChange({search: ''});
                 closeSearchPanel(this.view);
+                this.view.focus();
             }
 
             handleSearchNext() {
@@ -148,23 +142,18 @@ export const SearchPanelPlugin = (params: SearchPanelPluginParams) =>
                 findPrevious(this.view);
             }
 
-            handleSearchConfigChange(config: Partial<SearchQueryConfig>) {
-                this.setViewSearch(config);
-            }
-
-            handleReplaceNext(query: string, replacement: string) {
-                this.setViewSearch({search: query, replace: replacement});
+            handleReplaceNext() {
                 replaceNext(this.view);
             }
 
-            handleReplaceAll(query: string, replacement: string) {
-                this.setViewSearch({search: query, replace: replacement});
+            handleReplaceAll() {
                 replaceAll(this.view);
             }
         },
         {
             provide: () => [
                 keymap.of(searchKeymap),
+                searchTheme,
                 search({
                     createPanel: () => ({
                         // Create an empty search panel