+++ src/react-utils/useAutoFocus.ts
@@ -1,16 +1,19 @@
 import {type RefObject, useEffect} from 'react';
 
-export const useAutoFocus = (nodeRef: RefObject<HTMLElement>, dependencies: unknown[] = []) => {
+export function useAutoFocus<T extends HTMLElement>(
+    ref: RefObject<T>,
+    dependencies: unknown[] = [],
+    timeout = 0,
+) {
     useEffect(() => {
-        const {current: anchor} = nodeRef;
-        const timeout = setTimeout(() => {
-            anchor?.focus();
-        });
+        const timeoutId = setTimeout(() => {
+            ref.current?.focus();
+        }, timeout);
 
         return () => {
-            clearTimeout(timeout);
+            clearTimeout(timeoutId);
         };
         // https://github.com/facebook/react/issues/23392#issuecomment-1055610198
         // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [nodeRef.current, ...dependencies]);
-};
+    }, dependencies);
+}